+++ sample/sample9-custom-decorator/Logger.ts
@@ -4,6 +4,6 @@ import {ConsoleLogger} from "./ConsoleLogger";
 export function Logger() {
     return function(object: Object, propertyName: string, index?: number) {
         const logger = new ConsoleLogger();
-        Container.registerHandler({ object, propertyName, index, value: () => logger });
+        Container.registerHandler({ object, propertyName, index, value: containerInstance => logger });
     };
 }